Prepare R Markdown Template File

If the required template dependency files aren't present, download latest versions from the package website.

prepareTemplate(object, ...)

# S4 method for missing
prepareTemplate(object, sourceDir = NULL)

# S4 method for character
prepareTemplate(object, sourceDir = NULL)

Arguments

object
Optional. File name. If NULL (default), download the default dependency files for a new experiment.

...
Additional arguments (for the S4 generic definition).

sourceDir
Source directory, typically a URL, where the dependency files are located.

Value

No value.

Details

By default, this function will create local copies of these files:

_footer.Rmd

_header.Rmd

_output.yaml

bibliography.bib

setup.R

Examples

# Copy all of the default shared template files
prepareTemplate()
unlink(c(
    "_footer.Rmd",
    "_header.Rmd",
    "_output.yaml",
    "bibliography.bib",
    "setup.R"
))

# Request individual files
prepareTemplate("setup.R")
unlink("setup.R")

# Load the shared files from bcbioSingleCell
# NOT RUN {
prepareTemplate(
    sourceDir = system.file("rmarkdown/shared",
                            package = "bcbioSingleCell")
)
# }