Read Sample Metadata

This function reads user-defined sample metadata saved in a spreadsheet. The "description" column is always required, and must match the bcbio per sample directory names exactly. Inclusion of the "fileName" column isn't required but is recommended for data provenance.

readSampleData(file, lanes = 1L)

Arguments

file
string. File path. Supports CSV, TSV, and XLSX file formats.

lanes
scalar integer. Number of lanes used to split the samples into technical replicates (_LXXX) suffix.

Value

DataFrame.

Note

Some bcbio examples on readthedocs use "samplename" instead of "fileName". This function checks for that and will error out intentionally, since we're using the sampleName column (note case) to define unique sample names, in the event that bcbio has processed multiplexed samples.

Demultiplexed samples

This applies to bulk RNA-seq samples.

Normally when loading a bcbio run of demultiplexed samples, the sample metadata will be imported automatically from the project-summary.yaml file in the final upload directory using the readYAMLSampleData() function. If you notice any typos in your metadata after completing the run, these can be corrected by editing the YAML file. Alternatively, you can pass in a spreadsheet with the readSampleData() function.

The samples in the bcbio run must map to the description column. The values provided in description for demultiplexed samples must be unique. They must also be syntactically valid, meaning that they cannot contain illegal characters (e.g. spaces, non-alphanumerics, dashes) or begin with a number. Consult the documentation in help(topic = "make.names") for more information on valid names in R.

Multiplexed samples

This applies to some single-cell RNA-seq formats, including inDrops. In this case, bcbio will output per-sample directories with this this structure: description-revcomp.

readSampleData() checks to see if the description column is unique. If the values are duplicated, the function assumes that bcbio processed multiplexed FASTQs, where multiple samples of interest are barcoded inside a single FASTQ. This this case, you must supply additional "index", "sequence", and "sampleName" columns.

Note that bcbio currently outputs the reverse complement index sequence in the sample directory names (e.g. sample-ATAGAGAG). Define the forward index barcode in the sequence column here, not the reverse complement. The reverse complement will be calculated automatically and added as the revcomp column in the sample metadata.

Here's an example of valid multiplexed sample data:

| description | index | sequence | sampleName | genotype | |-------------|-------|----------|------------|----------| | indrops | 1 | CTCTCTAT | sample1 | wildtype | | indrops | 2 | TATCCTCT | sample2 | knockout | | indrops | 3 | GTAAGGAG | sample3 | wildtype | | indrops | 4 | ACTGCATA | sample4 | knockout |

See also

Examples

# Demultiplexed
file <- "http://bcbiobase.seq.cloud/demultiplexed.csv"
readr::read_csv(file)

#> Parsed with column specification:
#> cols(
#>   fileName = col_character(),
#>   description = col_character(),
#>   genotype = col_character()
#> )

#> # A tibble: 4 x 3
#>   fileName            description genotype
#>   <chr>               <chr>       <chr>   
#> 1 sample1_R1.fastq.gz sample1     wildtype
#> 2 sample2_R1.fastq.gz sample2     knockout
#> 3 sample3_R1.fastq.gz sample3     wildtype
#> 4 sample4_R1.fastq.gz sample4     knockout

x <- readSampleData(file)

#> Reading demultiplexed.csv

#> Demultiplexed samples detected

print(x)

#> DataFrame with 4 rows and 4 columns
#>         sampleName            fileName description genotype
#>           <factor>            <factor>    <factor> <factor>
#> sample1    sample1 sample1_R1.fastq.gz     sample1 wildtype
#> sample2    sample2 sample2_R1.fastq.gz     sample2 knockout
#> sample3    sample3 sample3_R1.fastq.gz     sample3 wildtype
#> sample4    sample4 sample4_R1.fastq.gz     sample4 knockout

# Multiplexed
file <- "http://bcbiobase.seq.cloud/multiplexed.csv"
readr::read_csv(file)

#> Parsed with column specification:
#> cols(
#>   fileName = col_character(),
#>   description = col_character(),
#>   index = col_integer(),
#>   sequence = col_character(),
#>   sampleName = col_character(),
#>   aggregate = col_character()
#> )

#> # A tibble: 8 x 6
#>   fileName             description index sequence sampleName aggregate
#>   <chr>                <chr>       <int> <chr>    <chr>      <chr>    
#> 1 indrops1_R1.fastq.gz indrops1        1 CTCTCTAT sample1_1  sample1  
#> 2 indrops1_R1.fastq.gz indrops1        2 TATCCTCT sample2_1  sample2  
#> 3 indrops1_R1.fastq.gz indrops1        3 GTAAGGAG sample3_1  sample3  
#> 4 indrops1_R1.fastq.gz indrops1        4 ACTGCATA sample4_1  sample4  
#> 5 indrops2_R1.fastq.gz indrops2        1 CTCTCTAT sample1_2  sample1  
#> 6 indrops2_R1.fastq.gz indrops2        2 TATCCTCT sample2_2  sample2  
#> 7 indrops2_R1.fastq.gz indrops2        3 GTAAGGAG sample3_2  sample3  
#> 8 indrops2_R1.fastq.gz indrops2        4 ACTGCATA sample4_2  sample4  

x <- readSampleData(file)

#> Reading multiplexed.csv

#> Multiplexed samples detected

print(x)

#> DataFrame with 8 rows and 7 columns
#>                   sampleName             fileName       description    index
#>                     <factor>             <factor>          <factor> <factor>
#> indrops1_AGAGGATA  sample2_1 indrops1_R1.fastq.gz indrops1-AGAGGATA        2
#> indrops1_ATAGAGAG  sample1_1 indrops1_R1.fastq.gz indrops1-ATAGAGAG        1
#> indrops1_CTCCTTAC  sample3_1 indrops1_R1.fastq.gz indrops1-CTCCTTAC        3
#> indrops1_TATGCAGT  sample4_1 indrops1_R1.fastq.gz indrops1-TATGCAGT        4
#> indrops2_AGAGGATA  sample2_2 indrops2_R1.fastq.gz indrops2-AGAGGATA        2
#> indrops2_ATAGAGAG  sample1_2 indrops2_R1.fastq.gz indrops2-ATAGAGAG        1
#> indrops2_CTCCTTAC  sample3_2 indrops2_R1.fastq.gz indrops2-CTCCTTAC        3
#> indrops2_TATGCAGT  sample4_2 indrops2_R1.fastq.gz indrops2-TATGCAGT        4
#>                   sequence aggregate  revcomp
#>                   <factor>  <factor> <factor>
#> indrops1_AGAGGATA TATCCTCT   sample2 AGAGGATA
#> indrops1_ATAGAGAG CTCTCTAT   sample1 ATAGAGAG
#> indrops1_CTCCTTAC GTAAGGAG   sample3 CTCCTTAC
#> indrops1_TATGCAGT ACTGCATA   sample4 TATGCAGT
#> indrops2_AGAGGATA TATCCTCT   sample2 AGAGGATA
#> indrops2_ATAGAGAG CTCTCTAT   sample1 ATAGAGAG
#> indrops2_CTCCTTAC GTAAGGAG   sample3 CTCCTTAC
#> indrops2_TATGCAGT ACTGCATA   sample4 TATGCAGT