Prepare Summarized Experiment
Source: R/prepareSummarizedExperiment.R

This is a utility wrapper for SummarizedExperiment::SummarizedExperiment() that provides automatic subsetting for row and column data.

prepareSummarizedExperiment(assays, rowRanges, colData, metadata = NULL,
  isSpike = NULL)

Arguments

assays
List containing RNA-seq count matrices with matching dimensions. Counts can be passed in either dense (matrix) or sparse (dgCMatrix, dgTMatrix) format.

rowRanges
GRanges describing assay rows. Must contain genomic ranges. Can be left NULL if the genome is poorly annotated and/or ranges aren't available from AnnotationHub.

colData
DataFrame data.frame, or matrix describing assay columns.

metadata
Optional. Metadata list.

isSpike
Character vector of spike-in sequence rownames.

Value

RangedSummarizedExperiment.

Details

This function also provides automatic metadata slotting of multiple useful environment parameters:

date: Today's date, returned from Sys.Date().

wd: Working directory, returned from getwd().

utilsSessionInfo: [utils::sessionInfo().

devtoolsSessionInfo: [sessioninfo::session_info()].

[utils::sessionInfo().

devtoolsSessionInfo: [sessioninfo::session_info()]: R:utils::sessionInfo().

Examples

genes <- c(
    "EGFP",  # spike
    "gene_1",
    "gene_2",
    "gene_3"
)
samples <- c(
    "sample_1",
    "sample_2",
    "sample_3",
    "sample_4"
)
mat <- matrix(
    seq(1L:16L),
    nrow = 4L,
    ncol = 4L,
    dimnames = list(genes, samples)
)
assays = list(assay = mat)
# Leave out the unannotated EGFP spike-in
rowRanges <- GRanges(
    seqnames = c("1", "1", "1"),
    ranges = IRanges(
        start = c(1L, 101L, 201L),
        end = c(100L, 200L, 300L)
    )
)
names(rowRanges) <- c("gene_1", "gene_2", "gene_3")
colData <- data.frame(
    "genotype" = c(
        "wildtype",
        "wildtype",
        "knockout",
        "knockout"
    ),
    "age" = c(3L, 6L, 3L, 6L),
    row.names = samples
)
prepareSummarizedExperiment(
    assays = assays,
    rowRanges = rowRanges,
    colData = colData,
    isSpike = "EGFP"
)

#> class: RangedSummarizedExperiment 
#> dim: 4 4 
#> metadata(6): date wd ... isSpike unannotatedRows
#> assays(1): assay
#> rownames(4): EGFP gene_1 gene_2 gene_3
#> rowData names(0):
#> colnames(4): sample_1 sample_2 sample_3 sample_4
#> colData names(2): genotype age